Redid how parts are connected to the spacecraft (see desc)
Parts now no longer use fixedJoints or have their own rigidbodies. They instead move under one spacecraft rigidbody. All parts also now have an adjustable mass as part of their scriptable object. Because of both of these changes, center of mass is now manually determined through the HandleSpacecraftMass() method in Spacecraft.cs.

In order to account for needing to click on objects during the build scene to drag them, parts start off with static rigidBodies, which are removed during flight. They are then readded as dyanmic rigidbodies when the spacecraft is destroyed so that they can separate and float freely.

Several changes were also made in a previous push that was labeled as "mid fix push". Some more changes will be pushed after this one, but they will just be for cleaning up the code.

+++ Assets/Scripts/PartScriptableObject.cs
@@ -6,7 +6,7 @@ public class PartScriptableObject : ScriptableObject {
 
     public int partID;
 
-    public int mass;
+    public float mass;
     
     //"This parts should only be connectable to the rest of the spacecraft from..."
     public string[] connectingDirections;
